Move duration, audio/title method to VideoPlayer.
