+++ CatchPlayAVKit/View/PlayerView.swift
@@ -9,16 +9,6 @@ protocol PlayerViewDelegate: AnyObject {
     func showPlayerControl(from playerView: PlayerView)
 }
 
-enum PlayerState {
-    case unknow
-    case readyToPlay
-    case playing
-    case buffering
-    case failed
-    case pause
-    case ended
-}
-
 import AVKit
 
 /// A view that displays the visual contents of a player object.
@@ -34,8 +24,6 @@ class PlayerView: UIView {
     }
     
     weak var delegate: PlayerViewDelegate?
-    
-    var playerState: PlayerState = .unknow
     
     private var playerLayer: AVPlayerLayer { layer as! AVPlayerLayer }
